+++ apps/web/src/types/comment/comment-service/feed-comment-service.d.ts
@@ -1,4 +1,4 @@
-import type { FeedReq, FeedUuid } from "../../feed-service";
+import type { FeedReq, FeedUuid } from "../../feed/common";
 import type {
   BaseComment,
   DeleteCommentReq,
