+++ apps/web/src/types/utility-service/bookmark-service.d.ts
@@ -1,5 +1,5 @@
 import type { CommonPaginationReq } from "../common";
-import type { FeedReq } from "../feed-service";
+import type { FeedReq } from "../feed/common";
 import type { ShortsReq } from "../shorts-service";
 
 export type GetShortsBookmarksReq = CommonPaginationReq;
